requirements.txt
bresenham==0.2.1
matplotlib==3.5.1
networkx==3.0
numpy==1.21.5
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.0.1
Pillow==9.5.0
